+++ tasks/image_processor/gaussianBlur.h
@@ -5,7 +5,7 @@ class Blur : public Filter {
 public:
     void ApplyFilter(Bmp& image, std::vector<double> args) override;
     ~Blur() override = default;
-    long double CountCoefficient(int x0, int x) const;
+    long double CountCoefficient(int64_t x0, int64_t x) const;
 
 private:
     double sigma_ = 1;
