+++ tasks/image_processor/matrixFilter.h
@@ -1,13 +1,14 @@
 #include "filter.h"
 #pragma once
-#include<vector>
+#include <vector>
 
 class MatrixFilter : public Filter {
 
 public:
     void ApplyFilter(Bmp& image, std::vector<double> args) override;
     ~MatrixFilter() override = default;
-    virtual int GetCoef(int i);
+    virtual int8_t GetCoef(size_t i);
+
 private:
-    std::vector<int> matrix_ = {0, 0, 0, 0, 0, 0, 0, 0, 0};
+    std::vector<int8_t> matrix_ = {0, 0, 0, 0, 0, 0, 0, 0, 0};
 };
